Use float constants.

+++ core/indigo-core/reaction/pathway_reaction.h
@@ -33,9 +33,9 @@ namespace indigo
 
     class DLLEXPORT PathwayReaction : public BaseReaction
     {
-        static constexpr int SPACE = 2;
-        static constexpr int ARROW_WIDTH = 3;
-        static constexpr int ARROW_MIN_HEIGHT = 5;
+        static constexpr float SPACE = 2.f;
+        static constexpr float ARROW_WIDTH = 3.f;
+        static constexpr float ARROW_MIN_HEIGHT = 5.f;
 
     public:
         PathwayReaction();
